scraper/requirements.txt
openai>=1.0.0
anthropic>=0.20.0
supabase>=2.0.0
python-dotenv>=1.0.0
